+++ experimental/navigation/utils/otel/spanCreator.ts
