+++ packages/core/scripts/__tests__/__mocks__/buildWithoutSwazzler.gradle
@@ -55,6 +55,8 @@ buildscript {
         
         
         
+        
+        
         classpath("com.facebook.react:react-native-gradle-plugin")
         classpath("de.undercouch:gradle-download-task:5.0.1")
         // NOTE: Do not place your application dependencies here; they belong
